+++ libs/solana/src/lib/helpers/index.ts
@@ -1,3 +1,4 @@
 export * from './get-public-key'
+export * from './parse-and-sign-token-transfer'
+export * from './parse-and-sign-transaction'
 export * from './parse-endpoint'
-export * from './parse-incoming-transaction'
